support for stop_gradient?
I had a case where I stop_gradients in my flow. Due to this, some of my grads are zero.

I've proposed a fix for this, but I should at the same time, ask if this invalidates the method or not, or if this is the correct way of handling no gradients?